requirements.txt
google-generativeai
pillow
ipython
exif
IPTCInfo3
piexif